+++ internal/admission/validator.go
@@ -284,7 +284,13 @@ func (validator KongHTTPValidator) ValidatePlugin(
 	var plugin kong.Plugin
 	plugin.Name = kong.String(k8sPlugin.PluginName)
 	var err error
-	plugin.Config, err = kongstate.RawConfigToConfiguration(k8sPlugin.Config)
+
+	plugin.Config, err = kongstate.RawConfigurationWithPatchesToConfiguration(
+		validator.SecretGetter,
+		k8sPlugin.Namespace,
+		k8sPlugin.Config,
+		k8sPlugin.ConfigPatches,
+	)
 	if err != nil {
 		return false, ErrTextPluginConfigInvalid, err
 	}
@@ -306,6 +312,11 @@ func (validator KongHTTPValidator) ValidatePlugin(
 	}
 	errText, err := validator.validatePluginAgainstGatewaySchema(ctx, plugin)
 	if err != nil || errText != "" {
+		validator.Logger.Info("validate KongPlugin on Kong gateway failed",
+			"plugin", fmt.Sprintf("%s/%s", k8sPlugin.Namespace, k8sPlugin.Name),
+			"config", plugin.Config,
+			"error", err,
+		)
 		return false, errText, err
 	}